remove typed-stack dependency
It was used for one purpose only: a stack abstraction in `except-args`.
Fortunately, the operations `top`, `stack-empty?` and `pop` (without
mutation) are represented by the purely functional `first`, `empty?`,
and `rest` over (linked) lists; we simulate mutation by carrying the
stack in the `let loop`.

This was found to be one of the heaviest remaining dependencies. At this
point, qi-lib only depends on "base" and "fancy-app," both of which are
necessary for the package to function as expected. It seems unlikely we
will get any more startup performance gains by trimming dependencies.