Change identifier-only behavior of `feedback` to "N"
The "while" behavior can always be expressed using
(feedback (while ...)) since the condition is a flow that has access
to the dynamic as well as the static context, whereas the (feedback N)
syntax only has access to the static context, and so it cannot use
runtime values. In the light of this, it is more useful for the
identifier form of `feedback` to exhibit the "N" semantics, as it
allows N to be derived from runtime values.