Fix feedback implementation to be able to use a binding
This is the same issue as we were seeing with partial application's
use of the `curry` form in its implementation, which required that the
arguments be available at compile time. We fixed it in the same way,
by wrapping the implementation in a lambda that accepts the runtime
arguments, allowing the use of bound identifiers in the feedback
specification.