Prioritize qi functions over racket functions in the expander
For unadorned identifiers to be treated as function identifiers,
ensure that qi functions take precedence over racket functions. This
allows us to define functions that may be treated as part of the
language (and not be shadowed by calling-scope identifiers) without
actually being syntactically part of the language as core forms or
macros.