Organize scripts to generate performance reports
Have separate scripts to generate reports of different aspects, to
preserve separation of concerns. This results in some duplication of
configuration, though, which is probably best avoided via some
improvements to the cli library.

Follow the general pattern of echoing live results to STDERR and
actual formatted output to STDOUT.