Add initial tests for the normalization pass
This uses the approach of independently normalizing two expressions we
expect to be equivalent, and comparing the results for equality. This
allows us to avoid dealing with the intricacies of the normalized
output in our tests while still making useful and sufficient
assertions about it.

The approach was suggested by Sam and Gus on Discourse:
https://racket.discourse.group/t/best-practices-for-testing-compiler-optimizations/2369