Simplify deforestation tests to high level assertions
This doesn't check that expressions are deforested _correctly_ so in
that respect it is worse than before, but it is in one respect more
useful and that is that it's possible to assert that deforestation is
_not_ happening without needing to know the exact expected target
expression.

If deforestation is not correctly done, we could expect that either
the regular unit tests or the benchmarks would reveal that.