Rewrite partial application to a use of a blanket template
This simplifies the core language by eliminating the need for
`#%partial-application` as a core form. It also allows us to deal with
the concept of chirality only in the expander and largely deal only in
terms of explicit templates in the compiler.

The only place where we still need chirality in the compiler is in
`clos`, which is a way to pre-supply arguments at runtime rather than
at compile time. There's no obvious syntactic way to designate which
side the pre-supplied arguments should be placed on here. We could
introduce `clos` and `closr` core forms to differentiate these cases
but that doesn't really simplify it (although it would be another step
in the direction of making our core language explicit and not reliant
on syntax properties).