Fix "anaphoric references" issue (resolves failing test)
See "The Artist Formerly Known as Bindingspec" in the Qi meeting notes
for more context on this issue. We had formerly fixed this for partial
application, but hadn't noticed the same issue also affected blanket
templates. Recently we committed a fix for the two cases of the
template being on the beginning or end (which is what partial
application without a template now expands to). This adds the fix for
the remaining case where a template is somewhere in the middle of the
expression.