Fix numeric `fanout` not getting optimized implementation
There is an optimized implementation for a literally indicated number
but that wasn't being used. It turns out it was because we hadn't
declared this in the Syntax Spec grammar, without which, it was
annotating the number with `#%host-expression` (which could have been
another way to fix this, by matching #%host-expression in the compiler).