Attempt to fix mutual inconsistency between `fix` and `find-and-map`
Both `fix` and `find-and-map` apply the same type of function (i.e. a
compiler rewrite rule) to syntax, but they have incompatible
expectations about the return value. Specifically, `fix` terminates on
a false return value, while `find-and-map` continues. This reconciles
them so that they both terminate upon receiving false, and both
continue if the transformed syntax is identical to the original.