Fix(?) weird syntax pair bug
`syntax-e` when given a syntax list seems to produce a syntax pair in
some cases. This was causing the `cdr` of the syntax _pair_ to
unexpectedly produce a syntax object which would then have the
transforming function inappropriately applied to it.

We don't know why syntax-e does this, but the fix for now is to first
attempt to use `syntax->list` which does more reliably produce a list,
and if that fails (e.g. identifiers), then use `syntax-e` as before.