Commit improved "weird syntax pair bug" fix from the meeting
This uses a new syntax property from Syntax Spec that identifies which
syntax subexpressions are core language forms, and these are the only
ones we will attempt to rewrite. We can't rely on either syntax-e or
syntax->list on its own since they aren't a robust source of this
information. For instance, [condition consequent] clauses of
`partition` would be picked up as lists, but it wouldn't be
appropriate to attempt to rewrite them.