First attempt at fixing the tests with the new syntax property
The tree traversal function `find-and-map/qi` now only attempts to
apply the transforming function if the `nonterminal` syntax property
is present on the syntax node, which indicates that it is a legitimate
use of a core Qi form, and thus one that it would be appropriate to
rewrite. The compiler tests are unaware of this new property so they
all fail. This modifies the syntax in the tests by first traversing it
and attaching the `nonterminal` property to every node, so that it
should be effectively the same as before the recent changes. But at the
moment, a couple of these tests are still failing.