Use on-demand expansion in normalization tests
This avoids the need to hand-write core language expressions, and
avoids the need to manually attach the `nonterminal` syntax
property. This also turns out to be more reliable since the
hand-written input expressions may not actually be encountered in
practice, whereas the ones produced by the expander from surface
syntax are guaranteed to be.