Introduce a `#%deforestable` core form
This form is intended to express any deforestable expression, allowing
the core language to express deforestation semantics, which, formerly,
we were not able to do within the language and thus resorted to
matching, and optimizing, host language syntax, leading to a "host" of
problems.

This new form is groundwork to enable compiler optimizations being
defined purely on the core language, thus representing a clean
boundary, or contract, between Qi and the host (Racket).

The initial implementation here just introduces the form, and code
generation for `filter` specifically, as a proof of concept for the
more generic and extensible planned implementation.

See the meeting notes for more, e.g.:
https://github.com/drym-org/qi/wiki/Qi-Meeting-Jun-21-2024#implementing-it