Add a commented-out failing test for a case of desired deforestation
When a nested form has a different chirality (threading direction)
than a containing form, normalization would not collapse them, but
deforestation may not care about the difference.

Possible approaches:

  A. Introduce normalization rules designed to detect
     when change of chirality is irrelevant.
  B. Look for patterns in the deforestation pass involving
    differing threading directions

Probably (A) is the right approach, and we could introduce a set of
chirality normalization rules that "trim" forms on either end of a
nested form which could be collapsed into the containing form. This
would include anything that isn't a host language function
application (which is the only case where chirality matters).

Actually, thinking again, chirality is already represented in the core
language simply as the presence of a blanket template in a function
application form, and nested threading is already collapsed by
normalization, so, I'm not sure anymore why this test is
failing ¯\_(ツ)_/¯