Integrate POC extension of deforestation to implement `map`
This integrates the proof-of-concept "deep macro" extension scheme for
deforestation that we have been working on in the weekly meetings. For
now, only `map` is implemented using the new approach, and though
user-level tests pass, some compiler tests are currently failing.

To keep things simple, this temporarily introduces a parallel
`#%deforestable2` form for this purpose so that we can implement list
operations one at a time and test things in isolation before going
whole hog and replacing `#%deforestable` with the new version.