Use the "clever hack" to translate `range`
We'd like to support multiple syntaxes for `range` that all expand to
a canonical form, and at the same time, provide a single codegen for
that canonical form. We don't already have a way to do this with
`define-deforestable` directly, so we just write `range` as an
ordinary Qi macro that expands to a use of the canonically-defined
`range2`.

See today's meeting notes for more on this.