Fully replace `#%deforestable` with the new implementation
This replaces the original implementation of the `#%deforestable` core
form with the new one that had been named `#%deforestable2` during the
process of incrementally integrating the proof-of-concept for this
scheme. With that complete, this now gets rid of the hardcoded codegen
for the `qi/list` forms in the Qi compiler, instead replacing it with
the ability to specify this in the `define-deforestable` macro which
then conveys the codegen through to the final stage of compilation via
a compile-time datatype.

This also removes some tests that are no longer relevant and updates
others.

The recent Qi meeting notes provide more context on this approach.