change deforestable "spec" syntax
Just `f` and `e` seem to make it hard to know which is the actual
binding position in, e.g., `[f f] [e e]`, so that we might be tempted
to write something like `[my-func f]`. Using the nonterminal names
`floe` and `expr` seem more explicit in indicating what kind of syntax
is expected while avoiding potential confusion with the binding
position.