Generalize `△` to exhibit `zip` behavior
It appears that △ and ▽ are closely related to the usual `zip` and
`unzip` operations in functional languages and represent a boundary
condition in their semantics.

Therefore, it's better to remove the tentative new `zip` core form in
favor of generalizing the existing △ form (and in the future,
likewise, also generalizing ▽).

This does entail a small backwards incompatibility, as `(△ flo)`
formerly accepted a single list and any number of additional values
that would be passed to each invocation of flo along with each element
of the list. That behavior is intuitively very similar to the "zip"
behavior, except that we now (more properly, it would seem) simply
expect these additional values to be present in separate input lists,
and these values may _happen_ to all be the same at each index.