bugfix: remove invalid error in `~>` re: `△`
As `△` now supports multiple input values, this error catching a
common mistake with a helpful message is unfortunately no longer
correct, as it also catches now-valid uses of `~>`.

Still, these multiple inputs are all expected to be _lists_, and it's
possible we can reintroduce a nice error in the event an input is not
a list. For now, this just removes the error handling to avoid the bug.