Only lift lets if any bindings are found
We found in today's meeting that an empty `let` was being added in the
expansion if there are no bindings, and the presence of this `let` was
enough to evade a CP0 optimization (which detects when there is a
single input argument and rewrites to avoid variadic args).

So we now only apply the "lifting lets" transformation in cases where
there actually are bindings in use.